Changed user's components
