+++ src/axios.js
@@ -2,8 +2,7 @@ import axios from "axios";
 import { toast } from "sonner";
 
 const axiosService = axios.create({
-  // baseURL: "https://double-decent-server.onrender.com/api/v1/",
-  baseURL: "http://localhost:5000/api/v1/",
+  baseURL: "https://double-decent-server.onrender.com/api/v1/",
   headers: { "Content-Type": "application/json" },
 });
