+++ src/components/checkout-success/CheckoutSuccess.jsx
@@ -1,4 +1,6 @@
+import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { Button } from "../ui/button";
 
 const CheckoutSuccess = () => {
     // Clear the cart after a successful order
@@ -63,18 +65,18 @@ const CheckoutSuccess = () => {
           </dl>
         </div>
         <div className="flex items-center space-x-4">
-          <a
+          <Button
             href="#"
-            className="text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
+            // className="text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
           >
             Track your order
-          </a>
-          <a
-            href="#"
-            className="py-2.5 px-5 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700"
+          </Button>
+          <Link
+            to="/products"
+            className="py-2.5 px-5 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700 hover:text-slate-700"
           >
             Return to shopping
-          </a>
+          </Link>
         </div>
       </div>
     </section>
