+++ src/pages/Orders/Orders.jsx
@@ -1,889 +1,191 @@
-import React from 'react'
+import { useState } from "react";
+import { useQuery } from "@tanstack/react-query";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import {
+  Table,
+  TableBody,
+  TableCaption,
+  TableCell,
+  TableFooter,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { Check, Ellipsis, Eye, Hourglass, RefreshCcw, Trash2, Truck, X } from "lucide-react";
+import axiosService from "@/axios";
+import { commaSeparatedPrice } from "@/utils/helperFunctions";
+
+import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
+import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
+import OrderDetails from "@/components/order-details/OrderDetails";
+import { Button } from "@/components/ui/button";
+
+import PaginationButton from "@/components/Pagination/Pagination";
 
 const Orders = () => {
-  return (
-    <div>
-      <section className="bg-white py-8 antialiased dark:bg-gray-900 md:py-16">
-        <div className="mx-auto max-w-screen-xl px-4 2xl:px-0">
-          <div className="mx-auto max-w-5xl">
-            <div className="gap-4 sm:flex sm:items-center sm:justify-between">
-              <h2 className="text-xl font-semibold text-gray-900 dark:text-white sm:text-2xl">
-                My orders
-              </h2>
+  const [activeTab, setActiveTab] = useState("all");
+  // Get all orders
+  const {
+    isLoading,
+    data: orders,
+    error,
+  } = useQuery({
+    queryKey: ["orders", activeTab],
+    queryFn: async () => {
+      const response = await axiosService.get(
+        `/order?type=${activeTab === "all" ? "" : activeTab}`
+      );
+      return response.data;
+    },
+  });
 
-              <div className="mt-6 gap-4 space-y-4 sm:mt-0 sm:flex sm:items-center sm:justify-end sm:space-y-0">
-                <div>
-                  <label
-                    htmlFor="order-type"
-                    className="sr-only mb-2 block text-sm font-medium text-gray-900 dark:text-white"
-                  >
-                    Select order type
-                  </label>
-                  <select
-                    id="order-type"
-                    className="block w-full min-w-[8rem] rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-primary-500 focus:ring-primary-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder:text-gray-400 dark:focus:border-primary-500 dark:focus:ring-primary-500"
-                  >
-                    <option selected>All orders</option>
-                    <option value="pre-order">Pre-order</option>
-                    <option value="transit">In transit</option>
-                    <option value="confirmed">Confirmed</option>
-                    <option value="cancelled">Cancelled</option>
-                  </select>
-                </div>
+  // console.log(activeTab, orders);
 
-                <span className="inline-block text-gray-500 dark:text-gray-400">
-                  {" "}
-                  from{" "}
-                </span>
+function formatDate(dateString) {
+  const date = new Date(dateString);
 
-                <div>
-                  <label
-                    htmlFor="duration"
-                    className="sr-only mb-2 block text-sm font-medium text-gray-900 dark:text-white"
-                  >
-                    Select duration
-                  </label>
-                  <select
-                    id="duration"
-                    className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-primary-500 focus:ring-primary-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder:text-gray-400 dark:focus:border-primary-500 dark:focus:ring-primary-500"
-                  >
-                    <option selected>this week</option>
-                    <option value="this month">this month</option>
-                    <option value="last 3 months">the last 3 months</option>
-                    <option value="lats 6 months">the last 6 months</option>
-                    <option value="this year">this year</option>
-                  </select>
-                </div>
-              </div>
-            </div>
+  // Extract the day, month, weekday, and year
+  const day = date.getDate(); // Day of the month
+  const month = date.toLocaleString("en-US", { month: "short" }); // Short month name (e.g., Jan, Feb)
+  const weekday = date.toLocaleString("en-US", { weekday: "short" }); // Short weekday name (e.g., Mon, Tue)
+  const year = date.getFullYear(); // Year
 
-            <div className="mt-6 flow-root sm:mt-8">
-              <div className="divide-y divide-gray-200 dark:divide-gray-700">
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB127364372
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      20.12.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $4,756
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-primary-100 px-2.5 py-0.5 text-xs font-medium text-primary-800 dark:bg-primary-900 dark:text-primary-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M18.5 4h-13m13 16h-13M8 20v-3.333a2 2 0 0 1 .4-1.2L10 12.6a1 1 0 0 0 0-1.2L8.4 8.533a2 2 0 0 1-.4-1.2V4h8v3.333a2 2 0 0 1-.4 1.2L13.957 11.4a1 1 0 0 0 0 1.2l1.643 2.867a2 2 0 0 1 .4 1.2V20H8Z"
-                        />
-                      </svg>
-                      Pre-order
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg border border-red-700 px-3 py-2 text-center text-sm font-medium text-red-700 hover:bg-red-700 hover:text-white focus:outline-none focus:ring-4 focus:ring-red-300 dark:border-red-500 dark:text-red-500 dark:hover:bg-red-600 dark:hover:text-white dark:focus:ring-red-900 lg:w-auto"
-                    >
-                      Cancel order
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB125467980
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      11.12.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $499
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-yellow-100 px-2.5 py-0.5 text-xs font-medium text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M13 7h6l2 4m-8-4v8m0-8V6a1 1 0 0 0-1-1H4a1 1 0 0 0-1 1v9h2m8 0H9m4 0h2m4 0h2v-4m0 0h-5m3.5 5.5a2.5 2.5 0 1 1-5 0 2.5 2.5 0 0 1 5 0Zm-10 0a2.5 2.5 0 1 1-5 0 2.5 2.5 0 0 1 5 0Z"
-                        />
-                      </svg>
-                      In transit
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg border border-red-700 px-3 py-2 text-center text-sm font-medium text-red-700 hover:bg-red-700 hover:text-white focus:outline-none focus:ring-4 focus:ring-red-300 dark:border-red-500 dark:text-red-500 dark:hover:bg-red-600 dark:hover:text-white dark:focus:ring-red-900 lg:w-auto"
-                    >
-                      Cancel order
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB139485607
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      08.12.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $85
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB137364371
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      16.11.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $119
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB134567890
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      02.11.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $2,056
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB146284623
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      26.09.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $180
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-red-100 px-2.5 py-0.5 text-xs font-medium text-red-800 dark:bg-red-900 dark:text-red-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M6 18 17.94 6M18 18 6.06 6"
-                        />
-                      </svg>
-                      Cancelled
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB145967376
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      17.07.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $756
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB148756352
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      30.06.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $235
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB159873546
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      04.06.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $90
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-red-100 px-2.5 py-0.5 text-xs font-medium text-red-800 dark:bg-red-900 dark:text-red-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M6 18 17.94 6M18 18 6.06 6"
-                        />
-                      </svg>
-                      Cancelled
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-
-                <div className="flex flex-wrap items-center gap-y-4 py-6">
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Order ID:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      <a href="#" className="hover:underline">
-                        #FWB156475937
-                      </a>
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Date:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      11.02.2023
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Price:
-                    </dt>
-                    <dd className="mt-1.5 text-base font-semibold text-gray-900 dark:text-white">
-                      $1,845
-                    </dd>
-                  </dl>
-
-                  <dl className="w-1/2 sm:w-1/4 lg:w-auto lg:flex-1">
-                    <dt className="text-base font-medium text-gray-500 dark:text-gray-400">
-                      Status:
-                    </dt>
-                    <dd className="me-2 mt-1.5 inline-flex items-center rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300">
-                      <svg
-                        className="me-1 h-3 w-3"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="24"
-                        height="24"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          stroke="currentColor"
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          stroke-width="2"
-                          d="M5 11.917 9.724 16.5 19 7.5"
-                        />
-                      </svg>
-                      Confirmed
-                    </dd>
-                  </dl>
-
-                  <div className="w-full grid sm:grid-cols-2 lg:flex lg:w-64 lg:items-center lg:justify-end gap-4">
-                    <button
-                      type="button"
-                      className="w-full rounded-lg bg-primary-700 px-3 py-2 text-sm font-medium text-white hover:bg-primary-800 focus:outline-none focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800 lg:w-auto"
-                    >
-                      Order again
-                    </button>
-                    <a
-                      href="#"
-                      className="w-full inline-flex justify-center rounded-lg  border border-gray-200 bg-white px-3 py-2 text-sm font-medium text-gray-900 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:outline-none focus:ring-4 focus:ring-gray-100 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:ring-gray-700 lg:w-auto"
-                    >
-                      View details
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            <nav
-              className="mt-6 flex items-center justify-center sm:mt-8"
-              aria-label="Page navigation example"
-            >
-              <ul className="flex h-8 items-center -space-x-px text-sm">
-                <li>
-                  <a
-                    href="#"
-                    className="ms-0 flex h-8 items-center justify-center rounded-s-lg border border-e-0 border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    <span className="sr-only">Previous</span>
-                    <svg
-                      className="h-4 w-4 rtl:rotate-180"
-                      aria-hidden="true"
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="24"
-                      height="24"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        stroke="currentColor"
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                        stroke-width="2"
-                        d="m15 19-7-7 7-7"
-                      />
-                    </svg>
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex h-8 items-center justify-center border border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    1
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex h-8 items-center justify-center border border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    2
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    aria-current="page"
-                    className="z-10 flex h-8 items-center justify-center border border-primary-300 bg-primary-50 px-3 leading-tight text-primary-600 hover:bg-primary-100 hover:text-primary-700 dark:border-gray-700 dark:bg-gray-700 dark:text-white"
-                  >
-                    3
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex h-8 items-center justify-center border border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    ...
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex h-8 items-center justify-center border border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    100
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="flex h-8 items-center justify-center rounded-e-lg border border-gray-300 bg-white px-3 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-                  >
-                    <span className="sr-only">Next</span>
-                    <svg
-                      className="h-4 w-4 rtl:rotate-180"
-                      aria-hidden="true"
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="24"
-                      height="24"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        stroke="currentColor"
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                        stroke-width="2"
-                        d="m9 5 7 7-7 7"
-                      />
-                    </svg>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </section>
-    </div>
-  );
+  // Combine the parts into the desired format
+  return `${day}, ${weekday} ${month} ${year}`;
 }
 
-export default Orders
+  if (isLoading) return <div>Loading...</div>;
+  if (error) return <div>Error: {error.message}</div>;
+  return (
+    <Tabs defaultValue="all" className="mx-1">
+      <TabsList className="grid w-full grid-cols-5 gap-1">
+        <TabsTrigger value="all">All</TabsTrigger>
+        <TabsTrigger value="paid" onClick={() => setActiveTab("paid")}>
+          Paid
+        </TabsTrigger>
+        <TabsTrigger
+          value="confirmed"
+          onClick={() => setActiveTab("confirmed")}
+        >
+          Confirmed
+        </TabsTrigger>
+        <TabsTrigger
+          value="delivered"
+          onClick={() => setActiveTab("delivered")}
+        >
+          Delivered
+        </TabsTrigger>
+        <TabsTrigger
+          value="cancelled"
+          onClick={() => setActiveTab("cancelled")}
+        >
+          Cancelled
+        </TabsTrigger>
+      </TabsList>
+      <TabsContent value={activeTab}>
+        <ScrollArea className="w-full whitespace-nowrap">
+          <Table>
+            <TableCaption>A list of your order history.</TableCaption>
+            <TableHeader>
+              <TableRow>
+                <TableHead>OrderID</TableHead>
+                <TableHead>Date</TableHead>
+                <TableHead>Price</TableHead>
+                <TableHead className="text-center">Status</TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {orders.length === 0 ? (
+                <div>No orders yet in {activeTab}</div>
+              ) : (
+                orders.map((order, index) => (
+                  <TableRow key={index}>
+                    <TableCell className="font-medium text-xs">
+                      {order._id}
+                    </TableCell>
+                    <TableCell>{formatDate(order.createdAt)}</TableCell>
+                    <TableCell>
+                      {commaSeparatedPrice(order.totalAmount)}
+                    </TableCell>
+                    <TableCell
+                      className={`capitalize w-fit m-auto rounded flex items-center ${
+                        order.status === "paid"
+                          ? "bg-blue-50 text-blue-500"
+                          : order.status === "cancelled"
+                          ? "bg-red-50 text-red-500"
+                          : order.status === "delivered"
+                          ? "bg-green-50 text-green-500"
+                          : "bg-yellow-50 text-yellow-500"
+                      } `}
+                    >
+                      {order.status === "paid" ? (
+                        <Hourglass size={15} />
+                      ) : order.status === "delivered" ? (
+                        <Check size={15} />
+                      ) : order.status === "cancelled" ? (
+                        <X />
+                      ) : (
+                        <Truck size={15} />
+                      )}
+                      <span className="ms-1">{order.status}</span>
+                    </TableCell>
+                    <TableCell className="text-right">
+                      <Dialog>
+                        <DropdownMenu>
+                          <DropdownMenuTrigger>
+                            <Ellipsis className="inline-block text-gray-500 cursor-pointer" />
+                          </DropdownMenuTrigger>
+                          <DropdownMenuContent className="p-2 me-2">
+                            <DropdownMenuItem className="font-semibold text-slate-600">
+                              <RefreshCcw /> <span>Order again</span>
+                            </DropdownMenuItem>
+                            <DropdownMenuItem className="font-semibold text-slate-600">
+                              <DialogTrigger className="flex items-center bg-transparent gap-1">
+                                <Eye /> <span>Order details</span>
+                              </DialogTrigger>
+                            </DropdownMenuItem>
+                            <DropdownMenuItem className="font-semibold">
+                              <Trash2 color="red" />
+                              <span className="text-red-500">Cancel order</span>
+                            </DropdownMenuItem>
+                          </DropdownMenuContent>
+                        </DropdownMenu>
+                        <DialogContent className="px-1 sm:p-6">
+                          <DialogHeader>
+                            <DialogTitle>Order details</DialogTitle>
+                            <DialogDescription>
+                              <ScrollArea className="h-[calc(100dvh-100px)]">
+                                <OrderDetails order={order} />
+                              </ScrollArea>
+                            </DialogDescription>
+                          </DialogHeader>
+                          <DialogFooter className="sm:justify-start">
+                            <DialogClose asChild>
+                              <Button type="button">Close</Button>
+                            </DialogClose>
+                          </DialogFooter>
+                        </DialogContent>
+                      </Dialog>
+                    </TableCell>
+                  </TableRow>
+                ))
+              )}
+            </TableBody>
+            {/* <TableFooter>
+                <TableRow>
+                  <TableCell colSpan={4} className="font-semibold">
+                    Total
+                  </TableCell>
+                  <TableCell className="text-right">$2,500.00</TableCell>
+                </TableRow>
+              </TableFooter> */}
+          </Table>
+          <ScrollBar orientation="horizontal" />
+        </ScrollArea>
+      </TabsContent>
+    </Tabs>
+  );
+};
+
+export default Orders;
